Remove executable permission for tests
Having executable bit set on tests means that nosetests will not find
the tests, so only 113 tests were running.  Correcting executable
permissions now yields 407 tests, all of which pass.

Original number of tests run by PyBuilder was 436.  This appears to be
due to duplicate function definitions in the test files