Modify tests for find_network_id
Tests for find_network_id were still using clc_sdk methods to find the
networks.

The following issue was remedied: when specifying a network id in the
Ansible playbook, if that network ID was not found, the method would
still return the first Network it found.  Now the code throws an error
to inform the user that it could not find the network that they were
looking for.