Update wait_until_complete behavior
If any of the count parameters were set to 0, the if checks were never
being triggered.  Now check whether the count parameters are not None.

Updated the _wait_until_complete behavior to keep polling until it gets
an explicit 'succeeded' or 'failed' status.

Due to the way _refresh_servers works, the objects returned are now
different.  Thus the behavior in start_stop_servers was returning the
old versions of the servers plus the updated versions.  Now combine two
lists before taking the set operation.

  - Fix logic based on counts, and fix data references
  - Fixed references to server.data['details']
  - Remove last references to requests module