Make firewall policy creation idempotent
This was initially raised in issue #13 and makes sense to implement in
the refactor of the ansible modules.  If firewall_policy_id is
specified, search through the existing policies to determine if a policy
exists that matches the parameters passed to the model.  If so, the
policy is not created.