Update loadbalancer module w/ native CLC API calls
The loadbalancer module is sychronous and blocking for all calls. As a
result, all calls need to wait to complete and require a timeout
generally longer than 30 seconds.
Additionally, the code flow was cleaned up so it is easier to follow the
steps that will be taken based on parameters passed to the module.