Update publicip module to use native API calls
In addition to using native API calls, also adjust how the public IPs
are assigned and created.  The old behavior was to create a new public
IP address on a new interface.  Now if no public IP address exists, by
default it gets added to the first internal IP address.  Additionally a
parameter was added where the internal IP address can be specified.